refactor: use sophia 0.8.0 [WIP]
